Make all UI components + only lib file TS
