+++ components/ui/VisualModeToggle.tsx
